+++ x-pack/plugins/lens/public/state_management/index.ts
@@ -8,11 +8,10 @@
 import { configureStore, getDefaultMiddleware, DeepPartial } from '@reduxjs/toolkit';
 import logger from 'redux-logger';
 import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux';
-import { lensSlice, initialState } from './lens_slice';
+import { lensSlice } from './lens_slice';
 import { timeRangeMiddleware } from './time_range_middleware';
 import { optimizingMiddleware } from './optimizing_middleware';
-import { LensAppState, LensState, StoreDeps } from './types';
-import { getInitialDatasourceId, getResolvedDateRange } from '../utils';
+import { LensState, LensStoreDeps } from './types';
 import { initMiddleware } from './init_middleware';
 export * from './types';
 
@@ -38,53 +37,9 @@ export const {
   setToggleFullscreen,
 } = lensSlice.actions;
 
-export const getPreloadedState = (
-  {
-    lensServices: { data },
-    initialContext,
-    embeddableEditorIncomingState,
-    datasourceMap,
-    visualizationMap,
-  }: StoreDeps,
-  preloadedState?: Partial<LensAppState>
-) => {
-  const initialDatasourceId = getInitialDatasourceId(datasourceMap);
-  const datasourceStates: LensAppState['datasourceStates'] = {};
-  if (initialDatasourceId) {
-    datasourceStates[initialDatasourceId] = {
-      state: null,
-      isLoading: true,
-    };
-  }
-
-  const state = {
-    lens: {
-      ...initialState,
-      isLoading: true,
-      query: data.query.queryString.getQuery(),
-      // Do not use app-specific filters from previous app,
-      // only if Lens was opened with the intention to visualize a field (e.g. coming from Discover)
-      filters: !initialContext
-        ? data.query.filterManager.getGlobalFilters()
-        : data.query.filterManager.getFilters(),
-      searchSessionId: data.search.session.getSessionId(),
-      resolvedDateRange: getResolvedDateRange(data.query.timefilter.timefilter),
-      isLinkedToOriginatingApp: Boolean(embeddableEditorIncomingState?.originatingApp),
-      activeDatasourceId: initialDatasourceId,
-      datasourceStates,
-      visualization: {
-        state: null as unknown,
-        activeId: Object.keys(visualizationMap)[0] || null,
-      },
-      ...preloadedState,
-    },
-  } as DeepPartial<LensState>;
-  return state;
-};
-
 export const makeConfigureStore = (
-  storeDeps: StoreDeps,
-  preloadedState?: Partial<LensAppState>
+  storeDeps: LensStoreDeps,
+  preloadedState: DeepPartial<LensState>
 ) => {
   const middleware = [
     ...getDefaultMiddleware({
@@ -99,7 +54,7 @@ export const makeConfigureStore = (
   return configureStore({
     reducer,
     middleware,
-    preloadedState: getPreloadedState(storeDeps, preloadedState),
+    preloadedState,
   });
 };
